+++ ta4j-core/src/main/java/org/ta4j/core/num/NaN.java
@@ -168,6 +168,11 @@ public class NaN implements Num {
     }
 
     @Override
+    public Num negate() {
+        return this;
+    }
+
+    @Override
     public boolean isZero() {
         return false;
     }